Minimal workaround for issue #7.
Attributes that the server sends, but the local API definition is
not aware of will be ignored. This should allow for some forward
compatibility as long as the new attributes are not required.

This issue stems from the design goal of minimizing dependencies
which eliminated the option of parsing the WSDL at runtime.

This is not an ideal solution - realistically, I would recommend
the M&M provided REST API now that it is available.